movement refactor + shoot wip
